requirements.txt
-i https://pypi.org/simple
alembic==1.12.0; python_version >= '3.7'
annotated-types==0.6.0; python_version >= '3.8'
anyio==3.7.1; python_version >= '3.7'
asyncpg==0.29.0; python_full_version >= '3.8.0'
cfgv==3.4.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
distlib==0.3.7
dnspython==2.4.2; python_version >= '3.8' and python_version < '4.0'
email-validator==2.1.0.post1
fastapi==0.104.1; python_version >= '3.8'
filelock==3.13.1; python_version >= '3.8'
greenlet==3.0.1; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
identify==2.5.31; python_version >= '3.8'
idna==3.4; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
mako==1.3.0; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
nodeenv==1.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
packaging==23.2; python_version >= '3.7'
passlib==1.7.4
platformdirs==3.11.0; python_version >= '3.7'
pluggy==1.3.0; python_version >= '3.8'
postgres==4.0
pre-commit==3.5.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
psycopg2-pool==1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pydantic[email]==2.4.2; python_version >= '3.7'
pydantic-core==2.10.1; python_version >= '3.7'
pydantic-settings==2.0.3; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
pytest==7.4.3; python_version >= '3.7'
pytest-asyncio==0.21.1; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
pyyaml==6.0.1; python_version >= '3.6'
setuptools==68.2.2; python_version >= '3.8'
sniffio==1.3.0; python_version >= '3.7'
sqlalchemy==2.0.20; python_version >= '3.7'
starlette==0.27.0; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
uvicorn==0.24.0.post1; python_version >= '3.8'
virtualenv==20.24.6; python_version >= '3.7'
